// src/lib/form-schemas/aditional-activity.schema.ts
import { z } from "zod"

export const aditionalActivitySchema = z.object({
	workBookId: z.number(),

	comments: z.string().optional(),
	executionDate: z.date({ message: "La fecha de ejecución no es válida" }),
	activityStartTime: z.string().nonempty({ message: "La hora de inicio no puede estar vacía" }),
	activityEndTime: z.string().nonempty({ message: "La hora de fin no puede estar vacía" }),
	activityName: z.string().nonempty({ message: "El nombre de la actividad no puede estar vacío" }),
})